Remove logging of successful configuration initialization
